天気データの取得をhooks化
