+++ src/components/detail/detail-weather.tsx
@@ -1,9 +1,9 @@
 import { Heading, Input } from '@/components/common'
-import { WeatherForecastType } from '@/types/weather'
 import { useEffect, useState } from 'react'
 import styled from 'styled-components'
 import { DateWeather } from '@/components/detail'
 import { useRouter } from 'next/router'
+import { useFetchWeather } from '@/hooks'
 
 const DetailWeatherContainer = styled.div`
   display: flex;
@@ -19,31 +19,11 @@ const DetailWeatherContainer = styled.div`
 `
 
 export const DetailWeather = () => {
-  const [weatherData, setWeatherData] = useState<
-    WeatherForecastType | undefined
-  >(undefined)
-  const [errorMessage, setErrorMessage] = useState('')
   const router = useRouter()
   const { query, isReady, push } = router
   const [queryDateValue, setQueryDateValue] = useState('')
   const [queryLocationValue, setQueryLocationValue] = useState('')
-
-  const fetchWeather = async (value: string) => {
-    try {
-      const response = await fetch(`/api/weather-forecast?location=${value}`)
-
-      if (!response.ok) {
-        throw new Error('取得失敗')
-      }
-
-      const data = (await response.json()) as WeatherForecastType
-      setWeatherData(data)
-    } catch (error) {
-      console.error(error)
-      setErrorMessage('天気データの取得に失敗しました')
-      setWeatherData(undefined)
-    }
-  }
+  const { fetchWeather, weatherData, errorMessage } = useFetchWeather()
 
   const handleNavigation = async (newLocation: string) => {
     await push(`/?location=${newLocation}`)
@@ -68,7 +48,7 @@ export const DetailWeather = () => {
     if (queryLocationValue) {
       void fetchWeather(queryLocationValue)
     }
-  }, [queryLocationValue, isReady, query])
+  }, [queryLocationValue, isReady, query, fetchWeather])
 
   const currentData =
     weatherData &&
